+++ lib/searchkick/query.rb
@@ -16,10 +16,10 @@ module Searchkick
 
     def initialize(klass, term = "*", **options)
       unknown_keywords = options.keys - [:aggs, :body, :body_options, :boost,
-        :boost_by, :boost_by_distance, :boost_where, :conversions, :conversions_term, :debug, :emoji, :exclude, :execute, :explain,
+        :boost_by, :boost_by_distance, :boost_by_recency, :boost_where, :conversions, :conversions_term, :debug, :emoji, :exclude, :execute, :explain,
         :fields, :highlight, :includes, :index_name, :indices_boost, :limit, :load,
         :match, :misspellings, :model_includes, :offset, :operator, :order, :padding, :page, :per_page, :profile,
-        :request_params, :routing, :scope_results, :select, :similar, :smart_aggs, :suggest, :track, :type, :where]
+        :request_params, :routing, :scope_results, :select, :similar, :smart_aggs, :suggest, :total_entries, :track, :type, :where]
       raise ArgumentError, "unknown keywords: #{unknown_keywords.join(", ")}" if unknown_keywords.any?
 
       term = term.to_s
@@ -111,10 +111,13 @@ module Searchkick
         model_includes: options[:model_includes],
         json: !@json.nil?,
         match_suffix: @match_suffix,
+        highlight: options[:highlight],
         highlighted_fields: @highlighted_fields || [],
         misspellings: @misspellings,
         term: term,
-        scope_results: options[:scope_results]
+        scope_results: options[:scope_results],
+        index_name: options[:index_name],
+        total_entries: options[:total_entries]
       }
 
       if options[:debug]
@@ -187,7 +190,7 @@ module Searchkick
         )
 
           raise UnsupportedVersionError, "This version of Searchkick requires Elasticsearch 5 or greater"
-        elsif e.message.include?("[multi_match] analyzer [searchkick_search] not found")
+        elsif e.message =~ /analyzer \[searchkick_.+\] not found/
           raise InvalidQueryError, "Bad mapping - run #{reindex_command}"
         else
           raise InvalidQueryError, e.message
@@ -224,9 +227,9 @@ module Searchkick
       @json = options[:body]
       if @json
         ignored_options = options.keys & [:aggs, :boost,
-          :boost_by, :boost_by_distance, :boost_where, :conversions, :conversions_term, :exclude, :explain,
-          :fields, :highlight, :indices_boost, :limit, :match, :misspellings, :offset, :operator, :order,
-          :padding, :page, :per_page, :profile, :select, :smart_aggs, :suggest, :where]
+          :boost_by, :boost_by_distance, :boost_by_recency, :boost_where, :conversions, :conversions_term, :exclude, :explain,
+          :fields, :highlight, :indices_boost, :match, :misspellings, :operator, :order,
+          :profile, :select, :smart_aggs, :suggest, :where]
         raise ArgumentError, "Options incompatible with body option: #{ignored_options.join(", ")}" if ignored_options.any?
         payload = @json
       else
@@ -248,7 +251,7 @@ module Searchkick
           if fields != ["_all"]
             query[:more_like_this][:fields] = fields
           end
-        elsif all
+        elsif all && !options[:exclude]
           query = {
             match_all: {}
           }
@@ -278,6 +281,15 @@ module Searchkick
             prefix_length = (misspellings.is_a?(Hash) && misspellings[:prefix_length]) || 0
             default_max_expansions = @misspellings_below ? 20 : 3
             max_expansions = (misspellings.is_a?(Hash) && misspellings[:max_expansions]) || default_max_expansions
+            misspellings_fields = misspellings.is_a?(Hash) && misspellings.key?(:fields) && misspellings[:fields].map(&:to_s)
+
+            if misspellings_fields
+              missing_fields = misspellings_fields - fields.map { |f| base_field(f) }
+              if missing_fields.any?
+                raise ArgumentError, "All fields in per-field misspellings must also be specified in fields option"
+              end
+            end
+
             @misspellings = true
           else
             @misspellings = false
@@ -312,8 +324,10 @@ module Searchkick
             exclude_analyzer = nil
             exclude_field = field
 
+            field_misspellings = misspellings && (!misspellings_fields || misspellings_fields.include?(base_field(field)))
+
             if field == "_all" || field.end_with?(".analyzed")
-              shared_options[:cutoff_frequency] = 0.001 unless operator.to_s == "and" || misspellings == false
+              shared_options[:cutoff_frequency] = 0.001 unless operator.to_s == "and" || field_misspellings == false
               qs << shared_options.merge(analyzer: "searchkick_search")
 
               # searchkick_search and searchkick_search2 are the same for ukrainian
@@ -332,7 +346,7 @@ module Searchkick
               exclude_analyzer = analyzer
             end
 
-            if misspellings != false && match_type == :match
+            if field_misspellings != false && match_type == :match
               qs.concat(qs.map { |q| q.except(:cutoff_frequency).merge(fuzziness: edit_distance, prefix_length: prefix_length, max_expansions: max_expansions, boost: factor).merge(transpositions) })
             end
 
@@ -372,21 +386,27 @@ module Searchkick
             end
           end
 
-          payload = {
-            dis_max: {
-              queries: queries
+          # all + exclude option
+          if all
+            query = {
+              match_all: {}
             }
-          }
 
-          should.concat(set_conversions)
+            should = []
+          else
+            payload = {
+              dis_max: {
+                queries: queries
+              }
+            }
+
+            should.concat(set_conversions)
+          end
 
           query = payload
         end
 
-        payload = {
-          size: per_page,
-          from: offset
-        }
+        payload = {}
 
         # type when inheritance
         where = (options[:where] || {}).dup
@@ -411,6 +431,7 @@ module Searchkick
         set_boost_by(multiply_filters, custom_filters)
         set_boost_where(custom_filters)
         set_boost_by_distance(custom_filters) if options[:boost_by_distance]
+        set_boost_by_recency(custom_filters) if options[:boost_by_recency]
 
         payload[:query] = build_query(query, filters, should, must_not, custom_filters, multiply_filters)
 
@@ -444,6 +465,13 @@ module Searchkick
         elsif load
           payload[:_source] = false
         end
+      end
+
+      # pagination
+      pagination_options = options[:page] || options[:limit] || options[:per_page] || options[:offset] || options[:padding]
+      if !options[:body] || pagination_options
+        payload[:size] = per_page
+        payload[:from] = offset
       end
 
       # type
@@ -482,9 +510,7 @@ module Searchkick
           ["_all"]
         elsif all && default_match == :phrase
           ["_all.phrase"]
-        elsif term == "*"
-          []
-        elsif default_match == :exact
+        elsif term != "*" && default_match == :exact
           raise ArgumentError, "Must specify fields to search"
         else
           [default_match == :word ? "*.analyzed" : "*.#{default_match}"]
@@ -579,12 +605,26 @@ module Searchkick
         unless attributes[:origin]
           raise ArgumentError, "boost_by_distance requires :origin"
         end
-        function_params = attributes.select { |k, _| [:origin, :scale, :offset, :decay].include?(k) }
+
+        function_params = attributes.except(:factor, :function)
         function_params[:origin] = location_value(function_params[:origin])
         custom_filters << {
           weight: attributes[:factor] || 1,
           attributes[:function] => {
             field => function_params
+          }
+        }
+      end
+    end
+
+    def set_boost_by_recency(custom_filters)
+      options[:boost_by_recency].each do |field, attributes|
+        attributes = {function: :gauss, origin: Time.now}.merge(attributes)
+
+        custom_filters << {
+          weight: attributes[:factor] || 1,
+          attributes[:function] => {
+            field => attributes.except(:factor, :function)
           }
         }
       end
@@ -622,11 +662,20 @@ module Searchkick
     def set_boost_by_indices(payload)
       return unless options[:indices_boost]
 
-      indices_boost = options[:indices_boost].each_with_object({}) do |(key, boost), memo|
-        index = key.respond_to?(:searchkick_index) ? key.searchkick_index.name : key
-        # try to use index explicitly instead of alias: https://github.com/elasticsearch/elasticsearch/issues/4756
-        index_by_alias = Searchkick.client.indices.get_alias(index: index).keys.first
-        memo[index_by_alias || index] = boost
+      if below52?
+        indices_boost = options[:indices_boost].each_with_object({}) do |(key, boost), memo|
+          index = key.respond_to?(:searchkick_index) ? key.searchkick_index.name : key
+          # try to use index explicitly instead of alias: https://github.com/elasticsearch/elasticsearch/issues/4756
+          index_by_alias = Searchkick.client.indices.get_alias(index: index).keys.first
+          memo[index_by_alias || index] = boost
+        end
+      else
+        # array format supports alias resolution
+        # https://github.com/elastic/elasticsearch/pull/21393
+        indices_boost = options[:indices_boost].map do |key, boost|
+          index = key.respond_to?(:searchkick_index) ? key.searchkick_index.name : key
+          {index => boost}
+        end
       end
 
       payload[:indices_boost] = indices_boost
@@ -699,7 +748,7 @@ module Searchkick
       aggs = Hash[aggs.map { |f| [f, {}] }] if aggs.is_a?(Array) # convert to more advanced syntax
       aggs.each do |field, agg_options|
         size = agg_options[:limit] ? agg_options[:limit] : 1_000
-        shared_agg_options = agg_options.slice(:order, :min_doc_count)
+        shared_agg_options = agg_options.except(:limit, :field, :ranges, :date_ranges, :where)
 
         if agg_options[:ranges]
           payload[:aggs][field] = {
@@ -716,19 +765,15 @@ module Searchkick
             }.merge(shared_agg_options)
           }
         elsif (histogram = agg_options[:date_histogram])
-          interval = histogram[:interval]
           payload[:aggs][field] = {
-            date_histogram: {
-              field: histogram[:field],
-              interval: interval
-            }
-          }
+            date_histogram: histogram
+          }.merge(shared_agg_options)
         elsif (metric = @@metric_aggs.find { |k| agg_options.has_key?(k) })
           payload[:aggs][field] = {
             metric => {
               field: agg_options[metric][:field] || field
             }
-          }
+          }.merge(shared_agg_options)
         else
           payload[:aggs][field] = {
             terms: {
@@ -778,7 +823,7 @@ module Searchkick
     # TODO id transformation for arrays
     def set_order(payload)
       order = options[:order].is_a?(Enumerable) ? options[:order] : {options[:order] => :asc}
-      id_field = :_uid
+      id_field = below60? ? :_uid : :_id
       payload[:sort] = order.is_a?(Array) ? order : Hash[order.map { |k, v| [k.to_s == "id" ? id_field : k, v] }]
     end
 
@@ -852,6 +897,8 @@ module Searchkick
                     }
                   }
                 }
+              when :prefix
+                filters << {prefix: {field => op_value}}
               when :regexp # support for regexp queries without using a regexp ruby object
                 filters << {regexp: {field => {value: op_value}}}
               when :not, :_not # not equal
@@ -902,7 +949,7 @@ module Searchkick
       elsif value.nil?
         {bool: {must_not: {exists: {field: field}}}}
       elsif value.is_a?(Regexp)
-        {regexp: {field => {value: value.source}}}
+        {regexp: {field => {value: value.source, flags: "NONE"}}}
       else
         {term: {field => value}}
       end
@@ -964,12 +1011,20 @@ module Searchkick
       end
     end
 
+    def base_field(k)
+      k.sub(/\.(analyzed|word_start|word_middle|word_end|text_start|text_middle|text_end|exact)\z/, "")
+    end
+
+    def below52?
+      Searchkick.server_below?("5.2.0")
+    end
+
     def below60?
-      Searchkick.server_below?("6.0.0-alpha1")
+      Searchkick.server_below?("6.0.0")
     end
 
     def below61?
-      Searchkick.server_below?("6.1.0-alpha1")
+      Searchkick.server_below?("6.1.0")
     end
   end
 end